Save sale button now returns a 2D list
