+++ SIMS/src/SIMS/RequestsPanel.java
@@ -37,8 +37,8 @@ public class RequestsPanel extends javax.swing.JPanel {
 		requestPanel = new javax.swing.JPanel();
 		helpButton = new javax.swing.JButton();
 		// Need to filter this to hide approved/denied orders
-		orderRequestPanel = new SIMS.OrderOrWasteRequestPanel("Order Requests", SupervisorWindow.getOrderTable());
-		wasteRequestPanel = new SIMS.OrderOrWasteRequestPanel("Waste Requests", SupervisorWindow.getWasteTable());
+		orderRequestPanel = new SIMS.OrderOrWasteRequestPanel("Order Requests", Database.getOrderTable());
+		wasteRequestPanel = new SIMS.OrderOrWasteRequestPanel("Waste Requests", Database.getWasteTable());
 
 		helpButton.setBackground(new java.awt.Color(255, 255, 153));
 		helpButton.setFont(new java.awt.Font("Segoe UI", 0, 24)); // NOI18N
